CICD/mlflow/requirements.txt
pandas
gunicorn
streamlit
scikit-learn==1.3.2  # Compatible avec Python 3.9
matplotlib
seaborn
plotly
mlflow
psycopg2-binary
xgboost==1.7.4
python-dotenv
pytest
requests==2.28.1
boto3
mock
imbalanced-learn
fsspec
s3fs
